LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/script-src-unsupported-type.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"

    <!-- speculative case in document.write -->
    <script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=50e634b6-be4d-46fb-80ee-d059827af01d&amp;encodingcheck=&Gbreve;" type=text/plain></script>


FAIL Speculative parsing, document.write(): script-src-unsupported-type Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"
